{Etwas bessere Doku, und Auftrennung des Codes.}
