+++ EclipseWS/Notos_Main/Localcode/States/LineFollow.cpp
@@ -6,6 +6,7 @@
  */
 
 #include "LineFollow.h"
+#include "Utility.h"
 
 namespace State {
 namespace LineFollow {
@@ -13,95 +14,27 @@ namespace LineFollow {
 using namespace Robot;
 using Communication::Pattern;
 
-void ramp() {
-	Led.setModes(Pattern::blink, 0, Pattern::blink);
-
-	Motor.setRotationSpeed(0);
-	Motor.setSpeed(MIN_MOTOR_SPEED);
-	while(Sensor::Tilting::isTilted()) {
-		_delay_ms(10);
-	}
-
-	Motor.moveBy(40);
-	Motor.flush();
-	Motor.continuousMode();
-
-}
-
-void buttonPause() {
-	if(getButton()) {
-		setMotors(false);
-
-		Led.setModes(0, 0, Pattern::slowblink);
-		_delay_ms(1000);
-
-		while(!getButton()) {
-			_delay_ms(100);
-		}
-		_delay_ms(1000);
-
-		setMotors(true);
-	}
-}
-
-void avoidObject() {
-	Led.setModes(0, Pattern::quickblink, Pattern::quickblink);
-
-	while(!getBumper()) {}
-
-	Motor.setSpeed(120);
-	Motor.setRotationSpeed(70);
-
-	Motor.moveBy(-50);
-	Motor.flush();
-	Motor.rotateBy(90);
-	Motor.flush();
-	Motor.moveBy(140);
-	Motor.flush();
-	Motor.rotateBy(-90);
-	Motor.flush();
-
-	Motor.moveBy(250);
-	Motor.flush();
-
-	Motor.rotateBy(-90);
-	Motor.flush();
-
-	Motor.setRotationSpeed(0);
-	Motor.setSpeed(100);
-	Motor.continuousMode();
-	while(LSensor.lineStatus != LF::Status::OK) {}
-
-	Motor.setRotationSpeed(100);
-	Motor.moveBy(50);
-	Motor.flush();
-	Motor.rotateBy(110);
-	Motor.flush();
-
-	Motor.continuousMode();
-}
-
 void simpleLF() {
 	Motor.continuousMode();
 
 	while(true) {
 		waitForSensors();
 
-		buttonPause();
+		Utility::pack();
 
-		if(getBumper())
-			avoidObject();
-		if(Sensor::Tilting::isTilted())
-			ramp();
-
+		// Normales Line-Following
 		if(LSensor.lineStatus == LF::Status::OK) {
-			Led.setModes(0, Pattern::on, 0);
-			Motor.setSpeed(fmax(
+			Led.setModes(0, Pattern::on, 0);		// Grünes Status-Licht
+			// Motorgeschwindigkeit "zweifach" proportional abhängig zwischen 0-LINE_CENTERED_CUTOFF und LCC-127
+			// Dies sorgt für High-Speed bei gerader Strecke, und feinen, problemlosen Kurven
+			Motor.setSpeed(fmax(					//
 			MID_MOTOR_SPEED - (MID_MOTOR_SPEED - MIN_MOTOR_SPEED) * fabs(LSensor.lineOffset)/127,
 			MAX_MOTOR_SPEED - (MAX_MOTOR_SPEED - MIN_MOTOR_SPEED) * fabs(LSensor.lineOffset)/LINE_CENTERED_CUTOFF));
 			
+			// Rotationsgeschwindigkeit proportional abhängig von der Auslenkung der Linie
 			Motor.setRotationSpeed(-((float)LSensor.lineOffset) * MAX_ROTATION_SPEED / 127);
 		}
+		// "Event"-Handling für das Verlieren der Linie
 		else {
 			if(LSensor.lineOffset < LINE_CENTERED_CUTOFF && LSensor.lineOffset > -LINE_CENTERED_CUTOFF) {
 				Led.setModes(0, 0, Pattern::blink);
