+++ EclipseWS/Notos_Main/Localcode/States/LineFollow.h
@@ -24,7 +24,6 @@ namespace State {
 namespace LineFollow {
 
 void simpleLF();
-void avoidObject();
 
 }
 }
